Change decoding error action
Also, include a new test case

Related to #9, but the test case does not quite cover it. Not sure what
the problem is.

We should really avoid broken encodings in the first place, but this
seems to be a direct output from hashcat. Some weird encoding of
the password perhaps?