Gracefully handle yet another encoding issue
Fixes #12.

If there is an encoding issue in the analytics report, hashcathelper
crashes with a stacktrace. This commit will cause it to print the
broken character instead, which is better than discarding the entire report.

The root of this is currently unknown, but appears to be the same as in #9.